Default value was returning long instead of int.
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@104292 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/jbaci/src/alma/ACS/impl/ROpatternImpl.java
@@ -106,7 +106,7 @@ public class ROpatternImpl
 	 * @see alma.ACS.PpatternOperations#default_value()
 	 */
 	public long default_value() {
-		return ((Long)defaultValue).longValue();
+		return ((Integer)defaultValue).intValue();
 	}
 
 	/**
